Write data to a database

Write data to a database, with options to insert new data, update existing data, and create a new table.

write_to_database(
  db,
  table_name,
  new_data,
  primary_keys,
  unique_indexes = NULL,
  insert_new = TRUE,
  update_duplicates = FALSE
)

Arguments

db

A database connection or path to database (if sqlite or duckdb file extension).

table_name

A character string specifying the table to write to.

new_data

A data frame to write to the table.

primary_keys

A character vector of column names to use as the primary key, the main identifier of individual rows in a table. Multiple columns can be specified and uniqueness will be assessed based on the combination of columns. (e.g. primary_keys = c("col1", "col2") will add a primary key on the combination of col1 and col2).

unique_indexes

A list of character vector(s) of column names to use as unique indexes. These will be added to the table, in addition to the primary key. Multiple columns can be specified and uniqueness will be assessed based on the combination of columns. (e.g. unique_indexes = list(c("col1", "col2")) will add a unique index on the combination of col1 and col2.). If NULL (the default), no unique indexes will be added.

insert_new

A logical value indicating if new data should be inserted into the existing table. If FALSE, no new data will be inserted, only existing rows will be updated if update_duplicates = TRUE.

update_duplicates

A logical value indicating if existing rows should be updated with new data. If FALSE, no existing rows will be updated. Only new data will be inserted if insert_new = TRUE.

Value

An invisible db connection.

Details

It will create the table and insert new_data if the table does not already exist. Otherwise:

If insert_new = TRUE, new_data will be inserted into the existing table.

If update_duplicates = TRUE, existing data will be updated with the overlap of new_data.